## Description
RDOQ for CHROMA

Closes #673

## Author

@hguermaz 

## Type of change

New feature/Tool

## Tests and performance
* BD-Rate to aom cpu0-2pass on the full list: ~0.6% Y gain and ~3.4% UV loss => YUV: ~0.075% loss (better distribution as currently UV to libaom is >> Y to libaom)
* Speed on a 4-core machine for M0 on 360p: ~1% slower
* Memory footprint on a 4-core machine: ~0%