@EwoutH 2019-10-08
No significant [performance impact](https://docs.google.com/spreadsheets/d/1p3PJQMkyhIrXEL6MBBQiwTnWmGyHY_zuw7_XAD0def4/edit#gid=43666624) as far as I could detect.

| FPS         | master |  PR667 | Increase |
|-------------|-------:|-------:|---------:|
| -enc-mode 8 | 61,560 | 61,598 |     0,1% |
| -enc-mode 7 | 30,430 | 30,438 |     0,0% |
| -enc-mode 6 | 22,057 | 22,131 |     0,3% |
| -enc-mode 5 | 13,726 | 13,688 |    -0,3% |
| -enc-mode 4 |  6,600 |  6,604 |     0,1% |
| -enc-mode 3 |  3,387 |  3,370 |    -0,5% |
| -enc-mode 2 |  1,574 |  1,563 |    -0,7% |
| -enc-mode 1 |  1,072 |  1,083 |     1,0% |
| -enc-mode 0 |  0,408 |  0,409 |     0,2% |